Simplify example scripts
